Deduplicate XLM/USD conversion helper
